Create isContractAddress util
